syncval: Segregate SyncOps class family into own file
