+++ BUILD.gn
@@ -239,9 +239,12 @@ vvl_sources = [
   "layers/stateless/stateless_validation.h",
   "layers/sync/sync_access_context.cpp",
   "layers/sync/sync_access_context.h",
+  "layers/sync/sync_common.cpp",
   "layers/sync/sync_common.h",
   "layers/sync/sync_model.cpp",
   "layers/sync/sync_model.h",
+  "layers/sync/sync_sync_op.cpp",
+  "layers/sync/sync_sync_op.h",
   "layers/sync/sync_utils.cpp",
   "layers/sync/sync_utils.h",
   "layers/sync/sync_validation.cpp",
